# internal/oas/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package oas

// Ref: #/components/schemas/CreateImage
type CreateImage struct {
	Title OptString `json:"title"`
	URL   string    `json:"url"`
}

// GetTitle returns the value of Title.
func (s *CreateImage) GetTitle() OptString {
	return s.Title
}

// GetURL returns the value of URL.
func (s *CreateImage) GetURL() string {
	return s.URL
}

// SetTitle sets the value of Title.
func (s *CreateImage) SetTitle(val OptString) {
	s.Title = val
}

// SetURL sets the value of URL.
func (s *CreateImage) SetURL(val string) {
	s.URL = val
}

// Ref: #/components/schemas/CreatePost
type CreatePost struct {
	Title    string    `json:"title"`
	Content  OptString `json:"content"`
	CoverURL OptString `json:"cover_url"`
}

// GetTitle returns the value of Title.
func (s *CreatePost) GetTitle() string {
	return s.Title
}

// GetContent returns the value of Content.
func (s *CreatePost) GetContent() OptString {
	return s.Content
}

// GetCoverURL returns the value of CoverURL.
func (s *CreatePost) GetCoverURL() OptString {
	return s.CoverURL
}

// SetTitle sets the value of Title.
func (s *CreatePost) SetTitle(val string) {
	s.Title = val
}

// SetContent sets the value of Content.
func (s *CreatePost) SetContent(val OptString) {
	s.Content = val
}

// SetCoverURL sets the value of CoverURL.
func (s *CreatePost) SetCoverURL(val OptString) {
	s.CoverURL = val
}

// Ref: #/components/schemas/Error
type Error struct {
	ErrorMessage string  `json:"error_message"`
	Code         float64 `json:"code"`
}

// GetErrorMessage returns the value of ErrorMessage.
func (s *Error) GetErrorMessage() string {
	return s.ErrorMessage
}

// GetCode returns the value of Code.
func (s *Error) GetCode() float64 {
	return s.Code
}

// SetErrorMessage sets the value of ErrorMessage.
func (s *Error) SetErrorMessage(val string) {
	s.ErrorMessage = val
}

// SetCode sets the value of Code.
func (s *Error) SetCode(val float64) {
	s.Code = val
}

func (*Error) imagesGetRes()      {}
func (*Error) imagesIDDeleteRes() {}
func (*Error) imagesIDGetRes()    {}
func (*Error) imagesIDPutRes()    {}
func (*Error) imagesPostRes()     {}
func (*Error) pingGetRes()        {}
func (*Error) postsGetRes()       {}
func (*Error) postsIDDeleteRes()  {}
func (*Error) postsIDGetRes()     {}
func (*Error) postsIDPutRes()     {}
func (*Error) postsPostRes()      {}

// Ref: #/components/schemas/Image
type Image struct {
	ID    int       `json:"id"`
	Title OptString `json:"title"`
	URL   string    `json:"url"`
}

// GetID returns the value of ID.
func (s *Image) GetID() int {
	return s.ID
}

// GetTitle returns the value of Title.
func (s *Image) GetTitle() OptString {
	return s.Title
}

// GetURL returns the value of URL.
func (s *Image) GetURL() string {
	return s.URL
}

// SetID sets the value of ID.
func (s *Image) SetID(val int) {
	s.ID = val
}

// SetTitle sets the value of Title.
func (s *Image) SetTitle(val OptString) {
	s.Title = val
}

// SetURL sets the value of URL.
func (s *Image) SetURL(val string) {
	s.URL = val
}

func (*Image) imagesIDGetRes() {}
func (*Image) imagesIDPutRes() {}

type ImagesGetOK struct {
	Data       []Image    `json:"data"`
	Pagination Pagination `json:"pagination"`
}

// GetData returns the value of Data.
func (s *ImagesGetOK) GetData() []Image {
	return s.Data
}

// GetPagination returns the value of Pagination.
func (s *ImagesGetOK) GetPagination() Pagination {
	return s.Pagination
}

// SetData sets the value of Data.
func (s *ImagesGetOK) SetData(val []Image) {
	s.Data = val
}

// SetPagination sets the value of Pagination.
func (s *ImagesGetOK) SetPagination(val Pagination) {
	s.Pagination = val
}

func (*ImagesGetOK) imagesGetRes() {}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// Ref: #/components/schemas/Pagination
type Pagination struct {
	CurrentPage  int `json:"currentPage"`
	TotalPages   int `json:"totalPages"`
	TotalItems   int `json:"totalItems"`
	ItemsPerPage int `json:"itemsPerPage"`
}

// GetCurrentPage returns the value of CurrentPage.
func (s *Pagination) GetCurrentPage() int {
	return s.CurrentPage
}

// GetTotalPages returns the value of TotalPages.
func (s *Pagination) GetTotalPages() int {
	return s.TotalPages
}

// GetTotalItems returns the value of TotalItems.
func (s *Pagination) GetTotalItems() int {
	return s.TotalItems
}

// GetItemsPerPage returns the value of ItemsPerPage.
func (s *Pagination) GetItemsPerPage() int {
	return s.ItemsPerPage
}

// SetCurrentPage sets the value of CurrentPage.
func (s *Pagination) SetCurrentPage(val int) {
	s.CurrentPage = val
}

// SetTotalPages sets the value of TotalPages.
func (s *Pagination) SetTotalPages(val int) {
	s.TotalPages = val
}

// SetTotalItems sets the value of TotalItems.
func (s *Pagination) SetTotalItems(val int) {
	s.TotalItems = val
}

// SetItemsPerPage sets the value of ItemsPerPage.
func (s *Pagination) SetItemsPerPage(val int) {
	s.ItemsPerPage = val
}

type PingGetOK struct{}

func (*PingGetOK) pingGetRes() {}

// Ref: #/components/schemas/Post
type Post struct {
	ID       int       `json:"id"`
	Title    string    `json:"title"`
	Content  OptString `json:"content"`
	CoverURL OptString `json:"cover_url"`
}

// GetID returns the value of ID.
func (s *Post) GetID() int {
	return s.ID
}

// GetTitle returns the value of Title.
func (s *Post) GetTitle() string {
	return s.Title
}

// GetContent returns the value of Content.
func (s *Post) GetContent() OptString {
	return s.Content
}

// GetCoverURL returns the value of CoverURL.
func (s *Post) GetCoverURL() OptString {
	return s.CoverURL
}

// SetID sets the value of ID.
func (s *Post) SetID(val int) {
	s.ID = val
}

// SetTitle sets the value of Title.
func (s *Post) SetTitle(val string) {
	s.Title = val
}

// SetContent sets the value of Content.
func (s *Post) SetContent(val OptString) {
	s.Content = val
}

// SetCoverURL sets the value of CoverURL.
func (s *Post) SetCoverURL(val OptString) {
	s.CoverURL = val
}

func (*Post) postsIDGetRes() {}
func (*Post) postsIDPutRes() {}

type PostsGetOK struct {
	Data       []Post     `json:"data"`
	Pagination Pagination `json:"pagination"`
}

// GetData returns the value of Data.
func (s *PostsGetOK) GetData() []Post {
	return s.Data
}

// GetPagination returns the value of Pagination.
func (s *PostsGetOK) GetPagination() Pagination {
	return s.Pagination
}

// SetData sets the value of Data.
func (s *PostsGetOK) SetData(val []Post) {
	s.Data = val
}

// SetPagination sets the value of Pagination.
func (s *PostsGetOK) SetPagination(val Pagination) {
	s.Pagination = val
}

func (*PostsGetOK) postsGetRes() {}

// Ref: #/components/schemas/Success
type Success struct {
	ID int `json:"id"`
}

// GetID returns the value of ID.
func (s *Success) GetID() int {
	return s.ID
}

// SetID sets the value of ID.
func (s *Success) SetID(val int) {
	s.ID = val
}

func (*Success) imagesIDDeleteRes() {}
func (*Success) imagesPostRes()     {}
func (*Success) postsIDDeleteRes()  {}
func (*Success) postsPostRes()      {}